feat(exceptions): introduce structured exception handling for policy SDK with detailed context
